refactor: move settings dialogues to settings/ directory
Move settings-dialogue, transcription-settings-dialogue, and
model-selector from src/components/ to src/components/settings/.

- Update imports in auth/user-settings.tsx, settings/model-selection.tsx, and project/project-main.tsx
- Export new components from settings/index.ts
- Fix implicitly any types in project-main.tsx
