+++ src/components/settings/index.ts
@@ -15,3 +15,6 @@ export { AdvancedReasoningSwitch } from "./advanced-reasoning-switch"
 export { SystemPromptInput } from "./system-prompt-input"
 export { AdvancedSettingsResetButton } from "./advanced-settings-reset-button"
 export { SubtitleCleanupSwitch } from "./subtitle-cleanup-switch"
+export { SettingsDialogue } from "./settings-dialogue"
+export { TranscriptionSettingsDialogue } from "./transcription-settings-dialogue"
+export { ModelSelector } from "./model-selector"
