+++ src/components/settings/transcription-settings-dialogue.tsx
@@ -12,7 +12,7 @@ import {
 import { Button } from "@/components/ui/button"
 import { Switch } from "@/components/ui/switch"
 import { Label } from "@/components/ui/label"
-import { SettingsTranscription } from "./transcribe/settings-transcription"
+import { SettingsTranscription } from "@/components/transcribe/settings-transcription"
 import { useTranscriptionDataStore } from "@/stores/data/use-transcription-data-store"
 import { GLOBAL_TRANSCRIPTION_SETTINGS_ID } from "@/constants/global-settings"
 import { DEFAULT_TRANSCRIPTION_SETTINGS } from "@/constants/default"
